+++ vsta/src/srv/tmpfs/stat.c
@@ -48,8 +48,8 @@ tmpfs_stat(struct msg *m, struct file *f)
 		len = o->o_len;
 		owner = o->o_owner;
 	}
-	sprintf(buf, "size=%d\ntype=%c\nowner=%d\ninode=%u\n",
-		len, f->f_file ? 'f' : 'd', owner, o);
+	sprintf(buf, "size=%d\ntype=%c\nowner=%d\ninode=%lu\n",
+		len, f->f_file ? 'f' : 'd', owner, (ulong)o);
 	if (o) {
 		strcat(buf, perm_print(&o->o_prot));
 	} else {
